requirements.txt
# app requirements
typer
ascii
regex
packaging
lxml[html-clean]
pyease-grpc>=1.6.0
python-dotenv>=0.15.0,<2.0.0
beautifulsoup4>=4.8.0,<5.0.0
requests>=2.20.0,<2.33.0
python-slugify>=4.0.0,<9.0.0
colorama>=0.4.0,<0.5.0
tqdm>=4.60,<5.0
PyExecJS>=1.5.1,<2.0.0
ebooklib>=0.17.0,<1.0.0
pillow>=6.0.0
cloudscraper>=1.2.71
readability-lxml>=0.8.0,<1.0.0
questionary>=1.6.0
prompt-toolkit~=3.0
html5lib~=1.1
base58~=2.1.1
python-box>=6.0.0,<8.0.0
pycryptodome>=3.0.0,<4.0.0
selenium>=3.141.0
tenacity>=9.0.0

# bot requirements
discord.py>=2.0.0
python-telegram-bot[job-queue]~=20.0
uvicorn
fastapi[standard]
cachetools
sqlmodel
passlib[argon2]
python-jose[cryptography]
python-dateutil
reflex

# dev requirements
wheel
black
flake8
tk-tools
setuptools
pyinstaller
types-tqdm
types-colorama
types-cachetools
types-python-dateutil
types-passlib
types-python-jose